pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dictation"
version = "0.1.0"
description = "Low-latency dictation pipeline: mic → ASR → cleanup LLM → text injection"
requires-python = ">=3.12"
dependencies = [
    "sounddevice>=0.5",
    "numpy>=1.26",
    "torch>=2.2",
    "torchaudio>=2.2",
    "pywhispercpp>=1.3",
    "pynput>=1.7",
    "pyobjc-framework-Quartz>=10.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "scipy>=1.12",
]

[project.scripts]
dictation = "dictation.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/dictation"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
